Unit Tests renamed into correct format
